[DUBBO-3137]: move constants from RegistryConstants back to dubbo-registry-api
