+++ dubbo-cluster/src/main/java/org/apache/dubbo/rpc/cluster/router/file/FileRouterFactory.java
@@ -27,7 +27,7 @@ import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 
-import static org.apache.dubbo.common.constants.RegistryConstants.ROUTER_KEY;
+import static org.apache.dubbo.rpc.cluster.Constants.ROUTER_KEY;
 import static org.apache.dubbo.rpc.cluster.Constants.RULE_KEY;
 import static org.apache.dubbo.rpc.cluster.Constants.RUNTIME_KEY;
 import static org.apache.dubbo.rpc.cluster.Constants.TYPE_KEY;
